Added AdvancedSettingsMixin
